+++ sonar-project.properties
@@ -4,7 +4,7 @@
 sonar.projectKey=open-cluster-management_klusterlet-addon-controller
 sonar.projectName=klusterlet-addon-controller
 sonar.sources=.
-sonar.exclusions=**/*_test.go,**/*_generated*.go,**/*_generated/**,**/vendor/**,**/vbh/**,./test/**,./pkg/apis/**,./pkg/dependencymagnet/**
+sonar.exclusions=**/*_test.go,**/*_generated*.go,**/*_generated/**,/test/**,/pkg/apis/**,/cmd/**,**/add_manager.go
 sonar.tests=.
 sonar.test.inclusions=**/*_test.go
 sonar.test.exclusions=**/*_generated*.go,**/*_generated/**,**/vendor/**,**/vbh/**,./test/**
